fix bug and oversight
In my original fix, PR'd yesterday, I had moved to change the position of the fridge to make it more friendly to the NCR  chemistry gameplay loop. Instead it was removed and not added because of my lack of being able to finish a project before jumping into another.

I'm sorry.

Here is the fix.