Legion Base Redecoration + More
Took two solid days to make it, but here we are. Legion base now has a story of what it used to be, and is mapped to reflect on that. Its now an old mining worksite, that's been occupied and built upon by the Legion to be their military camp. Has a mix of American and Legion buildings. Utilizes both z-levels. There is some general z-level realignment hotfixes, making ladders be on the same tile, the Oasis general store got a makeover, and I removed the religious posters and mistakenly placed posted in the Brotherhood base to put more appropriate ones down, along with some items that people placed inside walls.